Modifications to handle quests that may go for half days

+++ Assets/Code/Quests/IQSQuest.cs
@@ -13,6 +13,6 @@ namespace GG.QuestSystem
         string questReciever { set; get; }        //Hand in
                                                 //Item rewards
         int moneyReward { set; get; }           //gold rewards
-        int questTime { set; get; }             //Time Length - Maybe for AI parties only
+        float questTime { set; get; }             //Time Length - Maybe for AI parties only
     }
 }
